test for target directory
The code only used `target` "as-is" if it was a file.  This forces users who want to
compile a single file to a single target file to make sure that the target file exists.

This commit changes that to only write to a new file if the target is a directory.